requirements.txt
django==1.11
djangorestframework==3.6.2
mysql-connector==2.1.4
mysqlclient==1.3.10
djangorestframework-jwt==1.10
django-cors-headers==1.2.0
